day2
I'm glad I broke things out into functions like I did, it made modifying
it from part1 way easier as I just needed to change the logic in each of
the functions.

also, as I was reading it I totally predicted that it would change to
exactly this sort of thing, where the up and down wasn't an actual
movement but a slope sort of thing. love it