day 3 part 1
thing I ran across was that the sample had 5 digits of binary and the
actual file had more, so this code actually doesn't work with the sample
input as the starting point is wrong (since everything higher than
the 5th digit is zero, it'll skew the epsilon). I could fix that, but
lazy. I'm sure it'll come up in the second part though, so I'll probably
have to fix it there

also I could definitely DRY this out but again, lazy.

also I stole the stdin reading stuff from @petejohansen, it's so much
nicer than the horror show I had going on previously opening and
slurping in files.